[Keychron C1 Series] Reduce the amount of keymaps
