+++ keyboards/keychron/c1/keymaps/via_plain/keymap.c
@@ -1,3 +1,3 @@
 // Include default keymap
 // This way the VIA keymap is always up to date with the default keymap
-#include "../plain_ansi/keymap.c"
+#include "../default/keymap.c"
