+++ keyboards/keychron/c1/keymaps/via_white/rules.mk
@@ -18,4 +18,4 @@ OPT_DEFS += -DRAW_OUT_CAPACITY=1
 OPT_DEFS += -DCH_CFG_NO_IDLE_THREAD=TRUE
 
 # Include the patched functions
-SRC += keymaps/via/no_idle_thread.c
+SRC += no_idle_thread.c
